updated standard cookie cutter template
