+++ cookiecutter-android-example/{{ cookiecutter.repo_name }}/app/src/androidTest/java/github/nisrulz/example/{{ cookiecutter.package_path }}/ExampleInstrumentedTest.kt
