--- cookiecutter-android-example/{{ cookiecutter.repo_name }}/app/src/main/java/github/nisrulz/{{ cookiecutter.package_path }}/MainActivity.kt
@@ -1,12 +1,0 @@
-package {{ cookiecutter.full_package_namespace }}
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
