+++ cookiecutter-android-example/{{ cookiecutter.repo_name }}/app/src/test/java/github/nisrulz/example/{{ cookiecutter.package_path }}/ExampleUnitTest.kt
