+++ cookiecutter-android-example/{{ cookiecutter.repo_name }}/build.gradle
@@ -17,18 +17,10 @@ buildscript {
     }
 }
 
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-        
-    }
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
 
 subprojects {
     apply from: '../../dependencies.gradle'
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
 }
